reorganize new tests for multiple suspension points
